# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Portfolio Builder API",
    "description": "API for managing Themes, Users, Projects, and Skills with OAuth protected routes.",
    "version": "1.0.0"
  },

  "host": "cse41-final-project.onrender.com",
  "basePath": "/",
  "schemes": ["https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],

  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Use: Authorization: Bearer <token>"
    }
  },

  "tags": [
    { "name": "Themes", "description": "CRUD operations for themes" },
    { "name": "Users", "description": "CRUD operations for users" },
    { "name": "Projects", "description": "CRUD operations for projects" },
    { "name": "Skills", "description": "CRUD operations for skills" }
  ],

  "paths": {
    "/api/themes/": {
      "get": {
        "tags": ["Themes"],
        "summary": "Get all themes",
        "responses": {
          "200": {
            "description": "List of all themes",
            "schema": { "type": "array", "items": { "$ref": "#/definitions/Theme" } }
          },
          "500": { "description": "Internal Server Error" }
        }
      },
      "post": {
        "tags": ["Themes"],
        "summary": "Create a new theme (Protected)",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "body", "in": "body", "required": true, "schema": { "$ref": "#/definitions/NewTheme" } }
        ],
        "responses": {
          "201": { "description": "Theme created", "schema": { "$ref": "#/definitions/Theme" } },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },

    "/api/themes/{id}": {
      "get": {
        "tags": ["Themes"],
        "summary": "Get a theme by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }],
        "responses": {
          "200": { "description": "Theme found", "schema": { "$ref": "#/definitions/Theme" } },
          "400": { "description": "Invalid ID" },
          "404": { "description": "Theme not found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "put": {
        "tags": ["Themes"],
        "summary": "Update an existing theme (Protected)",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          { "name": "body", "in": "body", "required": true, "schema": { "$ref": "#/definitions/NewTheme" } }
        ],
        "responses": {
          "200": { "description": "Theme updated", "schema": { "$ref": "#/definitions/Theme" } },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Theme not found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "delete": {
        "tags": ["Themes"],
        "summary": "Delete a theme",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }],
        "responses": {
          "200": { "description": "Theme deleted" },
          "400": { "description": "Invalid ID" },
          "404": { "description": "Theme not found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },

    "/api/users/": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users",
        "responses": {
          "200": {
            "description": "List of all users",
            "schema": { "type": "array", "items": { "$ref": "#/definitions/User" } }
          },
          "500": { "description": "Internal Server Error" }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create a new user (Protected)",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "body", "in": "body", "required": true, "schema": { "$ref": "#/definitions/NewUser" } }
        ],
        "responses": {
          "201": { "description": "User created", "schema": { "$ref": "#/definitions/User" } },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },

    "/api/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get a user by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }],
        "responses": {
          "200": { "description": "User found", "schema": { "$ref": "#/definitions/User" } },
          "400": { "description": "Invalid ID" },
          "404": { "description": "User not found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update a user (Protected)",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          { "name": "body", "in": "body", "required": true, "schema": { "$ref": "#/definitions/NewUser" } }
        ],
        "responses": {
          "200": { "description": "User updated", "schema": { "$ref": "#/definitions/User" } },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete a user",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }],
        "responses": {
          "200": { "description": "User deleted" },
          "400": { "description": "Invalid ID" },
          "404": { "description": "User not found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },


    "/api/projects/": {
      "get": {
        "tags": ["Projects"],
        "summary": "Get all projects",
        "responses": {
          "200": {
            "description": "List of all projects",
            "schema": { "type": "array", "items": { "$ref": "#/definitions/Project" } }
          },
          "500": { "description": "Internal Server Error" }
        }
      },
      "post": {
        "tags": ["Projects"],
        "summary": "Create a new project (Protected)",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "body", "in": "body", "required": true, "schema": { "$ref": "#/definitions/NewProject" } }
        ],
        "responses": {
          "201": { "description": "Project created", "schema": { "$ref": "#/definitions/Project" } },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },

    "/api/projects/{id}": {
      "get": {
        "tags": ["Projects"],
        "summary": "Get a project by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }],
        "responses": {
          "200": { "description": "Project found", "schema": { "$ref": "#/definitions/Project" } },
          "400": { "description": "Invalid ID" },
          "404": { "description": "Project not found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "put": {
        "tags": ["Projects"],
        "summary": "Update a project (Protected)",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          { "name": "body", "in": "body", "required": true, "schema": { "$ref": "#/definitions/NewProject" } }
        ],
        "responses": {
          "200": { "description": "Project updated", "schema": { "$ref": "#/definitions/Project" } },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Project not found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "delete": {
        "tags": ["Projects"],
        "summary": "Delete a project",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }],
        "responses": {
          "200": { "description": "Project deleted" },
          "400": { "description": "Invalid ID" },
          "404": { "description": "Project not found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },

  
    "/api/skills/": {
      "get": {
        "tags": ["Skills"],
        "summary": "Get all skills",
        "responses": {
          "200": {
            "description": "List of all skills",
            "schema": { "type": "array", "items": { "$ref": "#/definitions/Skill" } }
          },
          "500": { "description": "Internal Server Error" }
        }
      },
      "post": {
        "tags": ["Skills"],
        "summary": "Create a new skill (Protected)",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "body", "in": "body", "required": true, "schema": { "$ref": "#/definitions/NewSkill" } }
        ],
        "responses": {
          "201": { "description": "Skill created", "schema": { "$ref": "#/definitions/Skill" } },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },

    "/api/skills/{id}": {
      "get": {
        "tags": ["Skills"],
        "summary": "Get a skill by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }],
        "responses": {
          "200": { "description": "Skill found", "schema": { "$ref": "#/definitions/Skill" } },
          "400": { "description": "Invalid ID" },
          "404": { "description": "Skill not found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "put": {
        "tags": ["Skills"],
        "summary": "Update a skill (Protected)",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          { "name": "body", "in": "body", "required": true, "schema": { "$ref": "#/definitions/NewSkill" } }
        ],
        "responses": {
          "200": { "description": "Skill updated", "schema": { "$ref": "#/definitions/Skill" } },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Skill not found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "delete": {
        "tags": ["Skills"],
        "summary": "Delete a skill",
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }],
        "responses": {
          "200": { "description": "Skill deleted" },
          "400": { "description": "Invalid ID" },
          "404": { "description": "Skill not found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    }
  },

  "definitions": {
    "Theme": {
      "type": "object",
      "properties": {
        "_id": { "type": "string" },
        "name": { "type": "string" },
        "description": { "type": "string" },
        "primaryColor": { "type": "string" },
        "secondaryColor": { "type": "string" },
        "fontFamily": { "type": "string" },
        "isDark": { "type": "boolean" }
      }
    },

    "NewTheme": {
      "type": "object",
      "required": ["name", "primaryColor", "secondaryColor", "fontFamily"],
      "properties": {
        "name": { "type": "string" },
        "description": { "type": "string" },
        "primaryColor": { "type": "string" },
        "secondaryColor": { "type": "string" },
        "fontFamily": { "type": "string" },
        "isDark": { "type": "boolean" }
      }
    },

    "User": {
      "type": "object",
      "properties": {
        "_id": { "type": "string" },
        "username": { "type": "string" },
        "email": { "type": "string" },
        "fullName": { "type": "string" },
        "bio": { "type": "string" },
        "avatarUrl": { "type": "string" }
      }
    },

    "NewUser": {
      "type": "object",
      "required": ["username", "email"],
      "properties": {
        "username": { "type": "string" },
        "email": { "type": "string" },
        "fullName": { "type": "string" },
        "bio": { "type": "string" },
        "avatarUrl": { "type": "string" }
      }
    },

    "Project": {
      "type": "object",
      "properties": {
        "_id": { "type": "string" },
        "title": { "type": "string" },
        "description": { "type": "string" },
        "imageUrl": { "type": "string" },
        "repoUrl": { "type": "string" },
        "liveUrl": { "type": "string" },
        "techStack": { "type": "string" }
      }
    },

    "NewProject": {
      "type": "object",
      "required": ["title"],
      "properties": {
        "title": { "type": "string" },
        "description": { "type": "string" },
        "imageUrl": { "type": "string" },
        "repoUrl": { "type": "string" },
        "liveUrl": { "type": "string" },
        "techStack": { "type": "string" }
      }
    },

    "Skill": {
      "type": "object",
      "properties": {
        "_id": { "type": "string" },
        "name": { "type": "string" },
        "level": { "type": "string" },
        "category": { "type": "string" },
        "iconUrl": { "type": "string" }
      }
    },

    "NewSkill": {
      "type": "object",
      "required": ["name"],
      "properties": {
        "name": { "type": "string" },
        "level": { "type": "string" },
        "category": { "type": "string" },
        "iconUrl": { "type": "string" }
      }
    }
  }
}